hacky treatment of ikiwiki wikilinks
rewrites them to be standard links to an id with name generated from
the textual form of the wikilink.

This is problematic in that the exact form of the
wikilink (e.g. relative vs absolute) will give a different resulting
name; some consistency will be required in authoring to make
everything work.  However, it works well enough to turn multiple
ikiwiki pages into a single document with internal links.