Updated LaTeX template:  `CJKoptions` variable, allow dvipsnames for colors.
Thanks to Xavier Olive.

The addition of the `CJKoptions` variable allows one to set (e.g.)
a bigger font size for Asian fonts than latin ones.

Including dvipsnames allows specifying colors like MidnightBlue
for link colors.  This brings in a dependency on the color package,
but it is a standard package and required anyway by graphics.