Removed obsolete reference to default.csl.
Closes #2372.

@cinaeco as the pandoc-citeproc man page describes, the current
behavior if no style is specified is to look in
`~/.csl/chicago-author-date.csl`, and if nothing is found there,
to use a default version of chicago-author-date.csl.

You may want to open a separate issue on jgm/pandoc-citeproc
suggesting the use of `default.csl` rather than
`chicago-author-date.csl`.  I agree that this would make sense.