round-tripping ikiwiki wikilinks
Since we now know (FSVO "know") what links resulted from wikilinks, we
can write them back out using ikiwiki wikilink syntax (assuming the
extension is turned on).

In order to get round-tripping to work, we no longer do any manipulation
of the link destination.  This is perhaps an unfortunate default, but
there's no other way to assure being able to round-trip.  In practice,
this means that filters will need to be used when exporting to other
formats in order to massage link destinations appropriately.