Realign LaTeX and beamer templates.
Adds new language and bidi variables; removes duplicated
`\VerbatimFootnotes`.

The standard `fontspec` variables, previously omitted, are now
included. Beamer defaults to the font set as `sansfont`. If `mainfont`
is set, the `serif` font theme will also be used, which uses this for
the slide text (does not have to be a serif font per se).

The paragraph indentation and use of the `url` package seem to have had
no visible effect, and have been removed.