Simplify hyperref usage.
The different usage for `xetex` appears to have been a workaround for bugs fixed in 2009 (see my question on this at <http://tex.stackexchange.com/questions/273339/what-hyperref-options-should-be-used-for-correct-unicode-output>), meaning that the same options can be used for all engines. The `subtitle` is used to set the PDF subject, following ConTeXt. The `bookmarks` option is enabled by default and does not need to be repeated.