Fixed wrong hyperref package syntax
As noted in [issue #174 on pandoc-citeproc](https://github.com/jgm/pandoc-citeproc/issues/174#issuecomment-151600792) internal links, only link to sections not to the actual target. This is due to the use of the wrong [hyperref package](http://ctan.org/pkg/hyperref) syntax. Internal links should use `hyperlink` and `hypertarget` instead of `hyperref` and `hyperdef`. This should fix the issue but is untested at the moment.