Docx writer:  better handling of PDF images.
Previously we tried to get the image size from the image even
if an explicit size was specified.  Since we still can't get
image size for PDFs, this made it impossible to use PDF images
in docx.

Now we don't try to get the image size when a size is already
explicitly specified.