Implement a Mallard Reader.
See http://projectmallard.org for more information, mallard is
basically a dumbed down docbook with the added notion of
http://projectmallard.org/1.0/mal_links , which usefulness
is very debatable.

The documentation maintainers of developer.gnome.org
might consider migrating to CommonMark, that's why I wrote this
reader, which is basically a copy pasted and trimmed down version
of the docbook one.

See https://github.com/GNOME/gnome-devel-docs for a decent corpus
of mallard pages.

The coverage is not total for this reader, but it's already useful
as is and I'll certainly get back to it pretty soon.