pandoc.hs: Implement parse-then-combine behavior.
Traditionally pandoc operates on multiple files by first concetenating
them (around extra line breaks) and then processing the joined
file. This has the benefit that footnotes and links can be in different
files, but it also introduces a couple of difficulties:

  - it is difficult to join files with footnotes without some sort of
    preprocessing, which makes it difficult to write academic documents
    in small pieces.

  - it makes it impossible to process multiple binary input files, which
    can't be catted.

  - it makes it impossible to process files from different input
    formats.

This commit introduces an experimental alternative method. Instead of
catting the files first, it parses the files first, and then combines
the parsed output. This makes it impossible to have links across
multiple files, and auto-identified headers won't work correctly if
headers in multiple files have the same name. On the other hand,
footnotes across multiple files will work correctly (THE major blocker
keeping me from using pandoc regularly for academic work) and will allow
more freedom for input formats.

In informal experiments, speed is roughly equivalent. In some
cases (where pandoc used to give warnings about duplicate footnotes) it
is faster.

Note that this is intended to be an optional method, defaulting to off.