pandoc.hs: parse first automatically if binary.
ByteStringReaders can currently only read one file, and will ignore
subsequent files. This changes the behavior to automatically parse
before combining if using the ByteStringReader. If we use one file, it
will work as normal. If there is more than one file it will combine them
after parsing (assuming that the format is the same).