Update LaTeX writer to align images
Update the LaTeX writer to align images that have an "align"
attribute whose value is "left", "right" or "center".  If the
attribute is absent or has an unrecognised values, the image is left
alone.

Also update the RST reader to read these horizonal alignment values
from the image directive (the first encountered value applies). See
http://docutils.sourceforge.net/docs/ref/rst/directives.html#image
for details.