Put only one heading element in top-level header
Every heading element semantically creates a new section.  Three
consecutive heading elements, e.g. `<h1></h1>`, `<h2></h2>`,
`<h3></h3>`, are semantically equivallent to a section with two nested
subsections.  This is not the intended meaning of subtitle, author, and
date, making plain `<p>` elements the better choice.  This is one of the
W3C's [common idioms].

This change is the result of discussions on issue jgm/pandoc#3119.

[common idioms]: https://www.w3.org/TR/html5/common-idioms.html#common-idioms

Fixes: jgm/pandoc#3119