Test case for not putting a dot '.' in id.
There is a problem when a title contains a . somewhere in the middle as
it gets into the generated id. That id cannot be matched by a CSS
selector as a dot marks a classname.

If you can give me any hints where to look for doing the fix. I'd be
glad to learn how.

Signed-off-by: Yoan Blanc <yoan@dosimple.ch>