LaTeX/Beamer writer:  allow hyperlinks to frames.
Previously you could link to a header above or below slide
level but not TO slide level.  This commit changes that.
Hypertargets are inserted inside frame titles; technically
the reference is to just after the title, but in normal
use (where slides are viewed full screen in a slide show),
this does not matter.

Closes #3220.